requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
numpy
pandas
tushare
python-dotenv
matplotlib